\name{likelihood_individual}
\alias{g3l_individual_linreg}
\alias{g3l_individual}
\concept{G3 action}

\title{Gadget3 likelihood actions for individual data}
\description{
  Compare model against a set of individual data points
}

\usage{
g3l_individual_linreg(
        fit = c('log', 'linear'),
        slope = 1,
        intercept = NULL )

g3l_individual(
        nll_name,
        obs_df,
        stocks,
        transform_f = quote( stock_ss(stock__wgt, vec = single) ),
        function_f = g3l_individual_linreg(),
        area_group = NULL,
        weight = g3_parameterized(paste0(nll_name, "_weight"),
            optimise = FALSE, value = 1),
        run_at = g3_action_order$likelihood )
}

\arguments{
  \item{slope, intercept}{
    \link{formula} substituted into surveyindices calcuations to fix slope/intercept of linear regression,
    or NULL if not fixed. See below.
  }
  \item{fit}{
   Is the fit 'log' or 'linear'?
  }
  \item{nll_name}{
    Character string, used to define the variable name for \var{obsstock} and \var{modelstock}.
  }
  \item{obs_df}{
    Data.frame of observation data.

    Should at least have a \var{year} & \var{step} column,
    some of \var{age}, \var{length} colunns & a \var{data} column for the observation.
  }
  \item{stocks}{
    A list of \code{\link{g3_stock}} objects to collect individual data for,
    depending if \var{stocks} were provided.
  }
  \item{transform_f}{
    \link{formula} to derive the model's equivalent predicted value for a data point.
  }
  \item{function_f}{
    A \link{formula} to compare \var{obs_df} to predicted values generated via \var{transform_f} and generate nll,
    defined by one of the \var{g3l_individual_}* functions.
  }
  \item{area_group}{
    List mapping area names used in \var{obs_df} to integer model areas,
    most likely generated by \code{\link{g3_areas}}.
  }
  \item{weight}{
    Weighting applied to this likelihood component. Default is a \code{g3_param}
    that defaults to 1, allowing weights to be altered without recompiling.
  }
  \item{run_at}{
    Integer order that actions will be run within model, see \code{\link{g3_action_order}}.
  }
}

\details{
  The actions will define the following variables in your model:
  \describe{
    \item{\var{nllstock}__obs}{Observation vector, the \var{data} column from \var{obs_df}}
    \item{\var{nllstock}__model}{The corresponding model prediction vector, total datapoints. \code{__model / __model_n} for the mean}
    \item{\var{nllstock}__model_n}{The number of data points at each point in the model prediction vector}
  }

  TODO:
}

\seealso{
  \url{https://gadget-framework.github.io/gadget2/userguide/chap-like.html},
  \code{\link{g3l_catchdistribution}}
  \code{\link{g3_stock}}
}

\value{
  \subsection{g3l_distribution_surveyindices_log}{
    Returns a \link{formula} for use as \var{function_f}:

    \deqn{
      \sum_{\it time} (\alpha + \beta \log N_{tral} - \log \frac{\nu_{tral}}{P_{tral}})^2
    }
    \describe{
      \item{\eqn{N_{tral}}}{Data column from \var{obs_df}}
      \item{\eqn{\nu_{tral}}}{Total predicted values \var{nllstock}__model}
      \item{\eqn{\P_{tral}}}{Number of data points \var{nllstock}__model_n}
      \item{\eqn{\alpha}}{\var{alpha} parameter}
      \item{\eqn{\beta}}{\var{beta} parameter}
    }

    If \var{alpha} or \var{beta} is not provided, then linear regression is
    performed on \eqn{N}, \eqn{\nu} over time for each area/age/length combination.
  }
}

\examples{
\dontshow{library(magrittr)}
st <- g3_stock("fish", c(10, 20, 30)) \%>\% g3s_age(3,5)

# Generate some random individual samples
obs_df <- data.frame(
    # NB: No 1993, we don't have any samples for that year
    year = rep(c(1990, 1991, 1992, 1994), each = 2),
    step = 1:2 )
obs_df$age = floor(runif(nrow(obs_df), min = 3, max = 5.1))
obs_df$length = floor(runif(nrow(obs_df), min = 10, max = 50))
obs_df$bt = runif(nrow(obs_df), min = 10, max = 1000)

actions <- list(
    g3a_time(1990, 1994, c(6,6)),
    # Use otherfood to populate abundance / mean weight
    g3a_otherfood(st,
        quote( age * 100 + stock__minlen ),
        quote( cur_year * 1e5 + cur_step * 1e4 + 0 * stock__minlen ) ),
    g3l_individual(
        "bt",
        obs_df,
        list(st),
        # TODO: Better example
        transform_f = g3_formula(
            stock_ss(stock__wgt, vec = single) + stock_ss(stock__num, vec = single),
            end = NULL ),
        function_f = g3l_individual_linreg(fit = "linear") ),
    NULL )

model_fn <- g3_to_r(c(actions, list(
    g3a_report_detail(actions),  # TODO: Not reporting anything useful
    NULL )))
r <- attributes(model_fn())

}
