\name{likelihood_bounds_penalty}
\alias{g3l_bounds_penalty}

\title{Gadget3 likelihood bounds_penalty action}
\description{
  Add a liklihood penalty for parameters leaving the bounds set in parameter_template
}

\usage{
g3l_bounds_penalty(
        parameter_template,
        weight = 1,
        run_at = g3_action_order$likelihood)
}

\arguments{
  \item{parameter_template}{
    A parameter template generated by code{\link{g3_to_tmb}}, with \var{optimise}, \var{lower}, \var{upper} populated.
  }
  \item{weight}{
    Weighting applied to this likelihood component.
  }
  \item{run_at}{
    Integer order that actions will be run within model, see "Ordering of actions" in \code{vignette("model_structure")}.
  }
}

\details{
  Whilst lower/upper can be passed to \code{\link{optim}}, not all methods can use them.
  Adding \code{g3l_bounds_penalty} OTOH can be used with any method.
}

\value{
  \subsection{g3l_distribution_bounds_penalty}{An action (i.e. list of formula objects) that will,
    for any parameter in \var{parameter_template}, add a large number to likelihood when outside the given bounds.
  }
}

\examples{
\dontshow{library(magrittr)}
anch <- g3_stock('anch', seq(20, 156, 4)) \%>\% g3s_age(3, 10)
actions <- list(
  g3a_time(1990, 1994),
  g3a_growmature(anch, g3a_grow_impl_bbinom(
    maxlengthgroupgrowth = 38L)),
  g3a_naturalmortality(anch),
  g3a_initialconditions_normalparam(anch),
  g3a_renewal_normalparam(anch,
    run_step = NULL),
  g3a_age(anch),
  NULL)

model_code <- g3_to_tmb(actions)
params.in <- attr(model_code, "parameter_template")

# Edit parameters to your liking...

# Regenerate model_code with bounds
model_code <- g3_to_tmb(c(actions, list(g3l_bounds_penalty(params.in))))
}